requirements.txt
matplotlib
numpy
tensorflow==2.11.1
tensorflow_datasets
flask
flask-cors
pillow